moved validateParameters into a new file with no build tag to use across linux and windows
